wasm_interp: create a test_utils module
